configure OnboardRouter in MapboxNavigation
